+++ bin/miden-cli/src/commands/init.rs
@@ -4,6 +4,7 @@ use clap::Parser;
 
 use crate::{
     config::{CliConfig, CliEndpoint},
+    errors::CliError,
     CLIENT_CONFIG_FILE_NAME,
 };
 
@@ -35,20 +36,22 @@ pub struct InitCmd {
 }
 
 impl InitCmd {
-    pub fn execute(&self, config_file_path: PathBuf) -> Result<(), String> {
+    pub fn execute(&self, config_file_path: PathBuf) -> Result<(), CliError> {
         if config_file_path.exists() {
-            return Err(format!(
-                "A filed named \"{}\" already exists in the working directory. Please try using another directory or removing the file.",
-                CLIENT_CONFIG_FILE_NAME
-            )
-            .to_string());
+            return Err(CliError::Config(
+                "Error with the configuration file".to_string(),
+                format!(
+                    "The file \"{}\" already exists in the working directory. Please try using another directory or removing the file.",
+                    CLIENT_CONFIG_FILE_NAME
+                ),
+            ));
         }
 
         let mut cli_config = CliConfig::default();
 
         if let Some(endpoint) = &self.rpc {
-            let endpoint =
-                CliEndpoint::try_from(endpoint.as_str()).map_err(|err| err.to_string())?;
+            let endpoint = CliEndpoint::try_from(endpoint.as_str())
+                .map_err(|err| CliError::Parse("Failed to parse RPC endpoint".to_string(), err))?;
 
             cli_config.rpc.endpoint = endpoint;
         }
@@ -62,18 +65,18 @@ impl InitCmd {
             None => None,
         };
 
-        let config_as_toml_string = toml::to_string_pretty(&cli_config)
-            .map_err(|err| format!("Error formatting config: {err}"))?;
+        let config_as_toml_string = toml::to_string_pretty(&cli_config).map_err(|err| {
+            CliError::Config("Failed to serialize config".to_string(), err.to_string())
+        })?;
 
-        let mut file_handle = File::options()
-            .write(true)
-            .create_new(true)
-            .open(&config_file_path)
-            .map_err(|err| format!("Error opening the file: {err}"))?;
+        let mut file_handle =
+            File::options().write(true).create_new(true).open(&config_file_path).map_err(
+                |err| CliError::Config("Failed to create config file".to_string(), err.to_string()),
+            )?;
 
-        file_handle
-            .write(config_as_toml_string.as_bytes())
-            .map_err(|err| format!("Error writing to file: {err}"))?;
+        file_handle.write(config_as_toml_string.as_bytes()).map_err(|err| {
+            CliError::Config("Failed to write config file".to_string(), err.to_string())
+        })?;
 
         println!("Config file successfully created at: {:?}", config_file_path);