+++ bin/miden-cli/src/main.rs
@@ -3,7 +3,7 @@ use miden_cli::Cli;
 extern crate std;
 
 #[tokio::main]
-async fn main() -> Result<(), String> {
+async fn main() -> miette::Result<()> {
     use clap::Parser;
 
     tracing_subscriber::fmt::init();
@@ -11,5 +11,5 @@ async fn main() -> Result<(), String> {
     let cli = Cli::parse();
 
     // execute cli action
-    cli.execute().await
+    Ok(cli.execute().await?)
 }